Add GPL headers. And misc fixes.

+++ src/whitehole/ObjectDBUpdater.java
@@ -1,10 +1,27 @@
+/*
+    Copyright 2012 Mega-Mario
+
+    This file is part of Whitehole.
+
+    Whitehole is free software: you can redistribute it and/or modify it under
+    the terms of the GNU General Public License as published by the Free
+    Software Foundation, either version 3 of the License, or (at your option)
+    any later version.
+
+    Whitehole is distributed in the hope that it will be useful, but WITHOUT ANY 
+    WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS 
+    FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License along 
+    with Whitehole. If not, see http://www.gnu.org/licenses/.
+*/
+
 package whitehole;
 
 import java.io.*;
 import java.net.*;
 import java.nio.ByteBuffer;
 import java.util.zip.CRC32;
-import java.nio.*;
 import java.nio.charset.*;
 import javax.swing.JLabel;
 
@@ -45,10 +62,17 @@ public class ObjectDBUpdater extends Thread
                 m_StatusLabel.setText("Object database already up-to-date.");
                 return;
             }
+            else if (data.length < 10)
+            {
+                m_StatusLabel.setText("Failed to update object database: received invalid data.");
+                return;
+            }
             
             CRC32 crc = new CRC32();
             crc.update(data, 9, data.length-9);
-            long crcref = Long.parseLong(strdata, 16);
+            long crcref;
+            try { crcref = Long.parseLong(strdata, 16); }
+            catch (NumberFormatException ex) { crcref = -1; }
             if (crc.getValue() != crcref)
             {
                 m_StatusLabel.setText("Failed to update object database: received invalid data.");
@@ -57,22 +81,24 @@ public class ObjectDBUpdater extends Thread
             
             File odbbkp = new File("objectdb.xml.bak");
             File odb = new File("objectdb.xml");
-            odb.renameTo(odbbkp);
-            try
+            if (odb.renameTo(odbbkp))
             {
-                odb.delete();
-                odb.createNewFile();
-                FileOutputStream odbstream = new FileOutputStream(odb);
-                odbstream.write(data, 9, data.length-9);
-                odbstream.flush();
-                odbstream.close();
-                odbbkp.delete();
-            }
-            catch (IOException ex)
-            {
-                m_StatusLabel.setText("Failed to save new object database.");
-                odbbkp.renameTo(odb);
-                return;
+                try
+                {
+                    odb.delete();
+                    odb.createNewFile();
+                    FileOutputStream odbstream = new FileOutputStream(odb);
+                    odbstream.write(data, 9, data.length-9);
+                    odbstream.flush();
+                    odbstream.close();
+                    odbbkp.delete();
+                }
+                catch (IOException ex)
+                {
+                    m_StatusLabel.setText("Failed to save new object database.");
+                    odbbkp.renameTo(odb);
+                    return;
+                }
             }
             
             m_StatusLabel.setText("Object database updated.");
